requirements.txt
langchain==0.0.154
openai==0.27.5
python-dotenv==1.0.0
fastapi==0.95.1
uvicorn[standard]==0.22.0
websockets==11.0.2
Jinja2==3.1.2
